requirements.txt
Django==1.11.11
djangorestframework==3.7.7
Pillow==5.0.0
pytz==2018.3
